+++ slb4j-ext/src/main/java/org/slb4j/ext/LogPane.java
@@ -15,10 +15,36 @@
  */
 package org.slb4j.ext;
 
+import org.slb4j.LogLayout;
+import org.slb4j.layout.PatternLayout;
+
 /**
  * Common interface for log pane components.
  */
 public interface LogPane {
+
+    /**
+     * The default pattern used for formatting log entries.
+     */
+    final LogLayout DEFAULT_LAYOUT = PatternLayout.parseLog4jPattern(
+            "%d{yyyy-MM-dd HH:mm:ss.SSS} [%t] %-5level %marker %logger{36} [%X]%n%n%msg%n%n%throwable%n"
+    );
+
+    /**
+     * Retrieves the current LogLayout instance used for formatting log entries in this LogPane's detail area.
+     *
+     * @return the LogLayout instance configuring the log formatting behavior
+     */
+    LogLayout getLogLayout();
+
+    /**
+     * Sets the LogLayout instance to be used for formatting log entries in this LogPane's detail area.
+     *
+     * @param layout the LogLayout instance that defines the formatting of log entries.
+     *               If null, the default layout may be applied or formatting may not occur.
+     */
+    void setLogLayout(LogLayout layout);
+
     /**
      * Retrieves the LogBuffer associated with this LogPane.
      *
@@ -40,4 +66,5 @@ public interface LogPane {
      * @return true if dark mode is enabled, otherwise false.
      */
     boolean isDarkMode();
+
 }